Set up most major methods, but trying to make them work
I added in a strip method to take out everything except single spaces
and words

I made a method that takes that string and puts it into an array

I made a selective sort method that rearranges the array into
alphabetical order, but I haven't been able to test if it works yet.

3/7    I went through and fixed some bugs.  Mark helped me figure out
that it would be good to use the split method to break apart the string
of text.  This works, but the method that I use to take out the spaces
as I transfer the information to an array list is not working.  I also
have a